pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pyphi"
version = "2.0.0a1"
description = "A package for integrated information theory"
readme = "README.md"
license = {text = "GNU General Public License v3 or later (GPLv3+)"}
authors = [
    {name = "William GP Mayner", email = "wmayner@gmail.com"}
]
keywords = [
    "neuroscience",
    "causality",
    "causal-modeling",
    "causation",
    "integrated-information-theory",
    "iit",
    "integrated-information",
    "modeling"
]
requires-python = ">=3.6"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Natural Language :: English",
    "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering"
]
dependencies = [
    "Graphillion>=1.5",
    "joblib>=0.8.0",
    "more_itertools>=8.13.0",
    "numpy>=1.11.0",
    "ordered-set>=4.0.2",
    "pandas>=2.0.0",
    "psutil>=2.1.1",
    "pyyaml>=3.13",
    "scipy>=0.13.3",
    "tblib>=1.3.2",
    "toolz>=0.9.0",
    "tqdm>=4.20.0",
]

[project.optional-dependencies]
dev = [
    "asv",
    "build",
    "coverage",
    "hypothesis",
    "pytest",
    "sphinx",
    "sphinx_rtd_theme",
    "twine",
    "watchdog"
]
parallel = [
    "ray[default]>=1.9.2"
]
visualize = [
    "matplotlib",
    "networkx>=2.6.2",
    "plotly>=5.8.2",
    "seaborn",
]
graphs = [
    "igraph>=0.9.10",
    "networkx>=2.6.2",
]
caching = [
    "redis>=2.10.5",
]
emd = [
    "pyemd>=0.3.0"
]

[project.urls]
documentation = "https://pyphi.readthedocs.io"
bugs = "https://github.com/wmayner/pyphi/issues"
homepage = "http://integratedinformationtheory.org/"
discussion = "https://groups.google.com/forum/#!forum/pyphi-users"

[tool.setuptools.packages.find]
exclude = ["docs", "test"]
